# apps/api-worker/openapi/worker-foundation.openapi.json
{
  "components": {
...
    "securitySchemes": {
      "SessionCookie": {
        "description": "Server-managed authenticated session cookie.",
        "in": "cookie",
        "name": "vocanova_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Operational VocaNova API running on Cloudflare Workers and D1.",
    "title": "VocaNova Worker API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/daily-mission": {
      "get": {
        "operationId": "GetDailyMission",
        "parameters": [
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "completedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "graceApplied": {
                      "type": "boolean"
                    },
                    "localDate": {
                      "format": "date",
                      "type": "string"
                    },
                    "newWordsCompleted": {
                      "type": "integer"
                    },
                    "newWordTarget": {
                      "type": "integer"
                    },
                    "policyVersion": {
                      "type": "string"
                    },
                    "reviewsCompleted": {
                      "type": "integer"
                    },
                    "reviewTarget": {
                      "type": "integer"
                    },
                    "sentencePracticesCompleted": {
                      "type": "integer"
                    },
                    "sentencePracticeTarget": {
                      "type": "integer"
                    },
                    "status": {
                      "enum": [
                        "open",
                        "completed",
                        "missed",
                        "protected"
                      ],
                      "type": "string"
                    },
                    "streak": {
                      "properties": {
                        "currentStreakCount": {
                          "type": "integer"
                        },
                        "graceDayBalance": {
                          "type": "integer"
                        },
                        "longestStreakCount": {
                          "type": "integer"
                        },
                        "status": {
                          "enum": [
                            "active",
                            "at_risk",
                            "broken"
                          ],
                          "type": "string"
                        }
                      },
                      "required": [
                        "currentStreakCount",
                        "longestStreakCount",
                        "status",
                        "graceDayBalance"
                      ],
                      "type": "object"
                    },
                    "timezone": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "localDate",
                    "timezone",
                    "reviewTarget",
                    "reviewsCompleted",
                    "policyVersion",
                    "status",
                    "graceApplied",
                    "streak"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Today's daily mission projection"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Missions"
        ]
      }
    },
...
    "/api/v1/progress": {
      "get": {
        "operationId": "GetProgress",
        "parameters": [
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "completionHistory": {
                      "items": {
                        "properties": {
                          "completed": {
                            "type": "boolean"
                          },
                          "localDate": {
                            "format": "date",
                            "type": "string"
                          }
                        },
                        "required": [
                          "localDate",
                          "completed"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "confidencePointsBalance": {
                      "type": "integer"
                    },
                    "streak": {
                      "properties": {
                        "currentStreakCount": {
                          "type": "integer"
                        },
                        "graceDayBalance": {
                          "type": "integer"
                        },
                        "longestStreakCount": {
                          "type": "integer"
                        },
                        "status": {
                          "enum": [
                            "active",
                            "at_risk",
                            "broken"
                          ],
                          "type": "string"
                        }
                      },
                      "required": [
                        "currentStreakCount",
                        "longestStreakCount",
                        "status",
                        "graceDayBalance"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "confidencePointsBalance",
                    "streak",
                    "completionHistory"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Progress projection"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Missions"
        ]
      }
    },
...
  },
...
}